Calculations with sets
There is an array of n integers. There are also 2 disjoint sets, A and B, each containing m integers. We like all the integers in set  A and dislike all the integers in set B. Our initial happiness is 0. For each i integer in the array, if i belongs A, we add 1 to our happiness. If i belongs B, we add  -1 to our happiness. Otherwise, our happiness does not change. Output our final happiness at the end.

Input Format

The first line contains integers n and m separated by a space. 
The second line contains n integers, the elements of the array. 
The third and fourth lines contain m integers, A and B, respectively.